chore: move tab components
