+++ packages/ui/index.ts
@@ -12,3 +12,4 @@ export { Input } from './Input';
 export { Modal } from './Modal';
 export { PageLoading } from './PageLoading';
 export { Spinner } from './Spinner';
+export { default as TabButton } from './TabButton';
